Change the class hierarchy of AbstractDualPi2PacketQueue (from which L4S and Classic queue classes derive) to AbstractPacketQueue, because: The L4S queue and the classic queue do not individually have limits. The limit is needed at the level of the DualQCoupledAQM, which manages these queues as a single pool.
Considering the above, having the AbstractDualPi2PacketQueue inherit from InfinitePacketQueue could have been an option, but this class does not accept arguments. Since the L4S queue does have arguments, this option is ruled out.
This leaves us with the AbstractPacketQueue option, with a copy of the useful members from DroppingPaquetQueue.